refactor(zhuyin): split polyphonicProcessor.ts into 4 pure modules (Related to #1843)
TDD-first: 44 characterization tests green before and after refactor.

Modules extracted:
- toneSandhi.ts: getNewToneForYiBu() + 一/不 Set constants
- polyphonicPatternMatcher.ts: match(), matchPattern(), PHRASES_ENDS_WITH_DI
- styleSetMapper.ts: variantIndexToStyleSet()
- zhuyinStringBuilder.ts: buildZhuyinString()

polyphonicProcessor.ts: 707L → 390L (singleton loader thin, delegates to modules)
buildZhuyinString re-exported for backward compat.
ZhuyinPhoneticGame.tsx and all consumers untouched.

Test coverage: tone sandhi (一/不), 地 de5/di4 phrases,
double-char pairs, PUA selector — all characterization locked.
