+++ frontend/src/components/zhuyin/polyphonicProcessor.ts
@@ -11,10 +11,12 @@
  *   - ButTaiwan/bpmfvs (bopomofo processing methods)
  */
 
-import { SS_MAPPING, ProcessedChar, PolyphonicData } from './bopomoConstants';
+import { ProcessedChar, PolyphonicData } from './bopomoConstants';
+import { matchPolyphonicPattern, isChineseCharacter } from './polyphonicPatternMatcher';
+import { variantIndexToStyleSet } from './styleSetMapper';
+import { getNewToneForYiBu } from './toneSandhi';
 import { getToneForChar } from './toneData';
-
-const CHINESE_CHAR_REGEX = /[\u4e00-\u9fa5]/;
+export { buildZhuyinString } from './zhuyinStringBuilder';
 
 /**
  * Singleton polyphonic processor.
@@ -104,306 +106,6 @@ export class PolyphonicProcessor {
       }
     }
     return filtered;
-  }
-
-  // ---------------------------------------------------------------------------
-  //  Tone sandhi for 一 and 不
-  // ---------------------------------------------------------------------------
-
-  /**
-   * 以下處理一、不的方法是基於 jeffreyxuan 的原始碼
-   * https://github.com/jeffreyxuan/toneoz-font-zhuyin/blob/main/src/js/ybtone.js
-   * 同時也請參考教育部國語辭典說明：https://dict.concised.moe.edu.tw/page.jsp?ID=55
-   *
-   * Returns [newSs, skipNext, skipPrev]
-   */
-  private getNewToneForYiBu(
-    prevChar: string | null,
-    currentChar: string,
-    nextChar: string | null,
-    prevTone: number | null,
-    nextTone: number | null,
-    skipPrev: boolean,
-  ): [string, boolean, boolean] {
-    const prevCharSet1 = new Set([
-      '第', '説', '说', '唯', '惟', '统', '統', '独', '獨', '劃', '划', '萬', '專', '某',
-      '十', '九', '八', '七', '六', '五', '四', '三', '二', '一', '〇', '零',
-    ]);
-
-    const nextCharSet1 = new Set([
-      '是', '日', '月', '的', '或', '物', '片', '系',
-      '十', '九', '八', '七', '六', '五', '四', '三', '二', '一', '〇', '零', '百', '千', '萬',
-      '元', '則', '節', '台', '同', '名', '回', '堂', '層', '幅', '幢', '年', '息', '成', '排',
-      '提', '搏', '擊', '擲', '旁', '時', '枚', '格', '條', '樓', '流', '環', '篇', '級', '群',
-      '言', '連', '門', '間',
-      '天', '經', '方', '對', '次', '家', '鳴', '命', '份', '件', '尊', '聲', '歲', '副', '本', '批',
-    ]);
-
-    // specialYiCases: nextChar → [newSs, skipNext, skipPrev]
-    const specialYiCases: Record<string, [string, boolean, boolean]> = {
-      '個': ['0000', false, false],
-      '个': ['0000', false, false],
-      '會': ['0000', false, false],
-      '切': ['0000', false, false],
-      '不': ['0000', true, true],
-    };
-
-    const specialBuCases: Record<string, [string, boolean, boolean]> = {
-      '禁': ['0000', false, false],
-      '菲': ['0000', false, false],
-      '勝': ['0000', false, false],
-      '著': ['0000', false, false],
-      '了': ['0000', false, false],
-      '好': ['0000', false, false],
-      '假': ['0000', false, false],
-      '當': ['ss01', false, false],
-    };
-
-    if (currentChar === '一') {
-      if (!skipPrev && (prevChar == null || prevCharSet1.has(prevChar))) {
-        return ['0000', false, true];
-      } else if (nextChar != null && nextChar in specialYiCases) {
-        return specialYiCases[nextChar]!;
-      } else if (nextChar == null || nextChar === '' || nextCharSet1.has(nextChar)) {
-        return ['0000', true, true];
-      } else if (prevChar === nextChar ||
-          ['看', '聽', '寫', '用', '說', '動', '搖', '問'].includes(nextChar ?? '')) {
-        return ['0000', true, true];
-      } else if (nextTone != null && (nextTone === 1 || nextTone === 2 || nextTone === 3)) {
-        return ['ss02', true, true]; // Fourth tone
-      } else if (nextTone != null && nextTone === 4) {
-        return ['ss01', true, true]; // Second tone
-      } else {
-        return ['0000', false, true];
-      }
-    } else if (currentChar === '不') {
-      if (nextChar != null && nextChar in specialBuCases) {
-        return specialBuCases[nextChar]!;
-      }
-      if (nextTone != null && (nextTone === 1 || nextTone === 2 || nextTone === 3 || nextTone === 5)) {
-        return ['0000', false, true]; // Remain 四聲
-      } else if (nextTone != null && nextTone === 4) {
-        return ['ss01', true, true]; // Change to second tone
-      } else {
-        return ['0000', false, true];
-      }
-    }
-
-    return ['0000', false, false]; // Default case
-  }
-
-  // ---------------------------------------------------------------------------
-  //  Pattern matching for polyphonic characters
-  // ---------------------------------------------------------------------------
-
-  /** Special set of three-char phrases ending with 地 that should read de5 */
-  private static readonly PHRASES_ENDS_WITH_DI = new Set([
-    '一十地', '大方地', '大聲地', '小心地', '小聲地', '不休地', '不安地',
-    '不倦地', '不停地', '不堪地', '不絕地', '不諱地', '不斷地', '亢奮地',
-    '仔細地', '叨叨地', '可憐地', '巧妙地', '平整地', '正當地', '正經地',
-    '生氣地', '生動地', '示弱地', '交替地', '吁吁地', '合適地', '吐吐地',
-    '如實地', '安靜地', '忙碌地', '成功地', '有味地', '有效地', '自主地',
-    '自由地', '自在地', '自信地', '自然地', '下氣地', '低聲地', '克難地',
-    '冷漠地', '吾吾地', '均勻地', '完整地', '忘我地', '快速地', '快樂地',
-    '抖擻地', '決然地', '牢固地', '狂暴地', '狂熱地', '甫定地', '迅速地',
-    '屈膝地', '周到地', '呱呱地', '和藹地', '坦率地', '委婉地', '怯步地',
-    '所能地', '易舉地', '虎嚥地', '采烈地', '勇敢地', '思索地', '急速地',
-    '筍般地', '耐心地', '重複地', '飛快地', '容易地', '恣意地', '悄悄地',
-    '特別地', '特定地', '真實地', '秘密地', '虔誠地', '究柢地', '高興地',
-    '乾脆地', '停蹄地', '偷偷地', '堅定地', '堅強地', '堅毅地', '專注地',
-    '康康地', '強烈地', '得意地', '悠悠地', '悠揚地', '悠閒地', '情願地',
-    '授權地', '敏感地', '敏銳地', '淡寫地', '深刻地', '深深地', '清楚地',
-    '甜甜地', '細心地', '許可地', '尊敬地', '悲傷地', '惺忪地', '愉快地',
-    '無私地', '無償地', '猶豫地', '痛苦地', '絮絮地', '間斷地', '意外地',
-    '意料地', '準確地', '溫柔地', '煞氣地', '痴痴地', '經意地', '詳細地',
-    '誠意地', '誠懇地', '嘆氣地', '慢慢地', '慣性地', '漂亮地', '漸漸地',
-    '瘋狂地', '盡力地', '盡瘁地', '緊緊地', '輕盈地', '輕微地', '輕輕地',
-    '輕聲地', '遠遠地', '嘩啦地', '熟慮地', '熟練地', '熱心地', '熱情地',
-    '範圍地', '緩慢地', '緩緩地', '踏實地', '整齊地', '激動地', '興奮地',
-    '諱言地', '錯誤地', '隨意地', '靜靜地', '靦腆地', '默默地', '優雅地',
-    '翼翼地', '闊步地', '禮貌地', '簡單地', '謹慎地', '穩固地', '穩穩地',
-    '嚴厲地', '歡快地', '驕傲地', '驚恐地', '靈活地', '究底地', '大大地',
-  ]);
-
-  /**
-   * Match polyphonic patterns for a character.
-   * Returns [matchIndex, skipNext, skipPrev]
-   */
-  private match(
-    character: string,
-    index: number,
-    patterns: string[],
-    text: string[],
-    skipPrev: boolean = false,
-  ): [number, boolean, boolean] {
-    let defaultIndex = -1;
-    const prev2Char = index > 1 ? text[index - 2] : '';
-    const prevChar = index > 0 ? text[index - 1] : '';
-    const threeCharPhrase = prev2Char + prevChar + character;
-    const nextChar = index + 1 < text.length ? text[index + 1] : '';
-    const isFirstChar = prevChar === '' || !this.isChineseCharacter(prevChar);
-    const isLastChar = nextChar === '' || !this.isChineseCharacter(nextChar);
-    const isStandalone = text.length === 1 || (isFirstChar && isLastChar);
-    let skipNext = false;
-
-    if (isStandalone) {
-      return [0, false, true];
-    }
-
-    // Special handle for '地' with 'de5' sound
-    if (character === '地') {
-      if (PolyphonicProcessor.PHRASES_ENDS_WITH_DI.has(threeCharPhrase)) {
-        return [1, false, true]; // de5
-      } else {
-        return [0, false, true]; // di4
-      }
-    }
-
-    // Special handle for '著' (著作權)
-    if (character === '著') {
-      const secondLine = patterns.length > 1 ? patterns[1] : '';
-      if (index >= 0 && index + 2 < text.length) {
-        const prefix = character + text[index + 1] + text[index + 2];
-        if (prefix === '著作權') {
-          return [patterns.indexOf(secondLine), false, false];
-        }
-      }
-    }
-
-    // First pass: Checking all patterns for "any+*" or "any+*+any"
-    if (!isFirstChar && !skipPrev) {
-      for (let j = 0; j < patterns.length; j++) {
-        const combinedPattern = patterns[j];
-        const subPatterns = combinedPattern.split('/');
-
-        for (const pattern of subPatterns) {
-          if (pattern === '') {
-            defaultIndex = j;
-            continue;
-          }
-          if (pattern.startsWith('*')) continue; // Skip patterns starting with '*' in first pass
-          const pos = pattern.indexOf('*');
-          if (pos === -1) continue;
-
-          const start = index - pos;
-          const end = index - pos + pattern.length;
-          if (start < 0 || end > text.length) continue;
-
-          if (this.matchPattern(pattern, pos, start, end, text, character)) {
-            return [j, false, true];
-          }
-        }
-      }
-    }
-
-    // Second pass: Checking patterns for "*+any"
-    if (!isLastChar) {
-      for (let j = 0; j < patterns.length; j++) {
-        const combinedPattern = patterns[j];
-        const subPatterns = combinedPattern.split('/');
-
-        for (const pattern of subPatterns) {
-          if (pattern === '') {
-            defaultIndex = j;
-            continue;
-          }
-          if (!pattern.startsWith('*')) continue; // Only patterns starting with '*' in second pass
-          const pos = pattern.indexOf('*');
-          if (pos === -1) continue;
-
-          const start = index - pos;
-          const end = index - pos + pattern.length;
-          if (start < 0 || end > text.length) continue;
-
-          if (this.matchPattern(pattern, pos, start, end, text, character)) {
-            if (this.isPolyphonicChar(nextChar)) {
-              skipNext = false;
-              skipPrev = false;
-            } else {
-              skipPrev = true;
-              skipNext = pattern.startsWith('*') && pos === 0;
-            }
-            if (this.isPolyphonicChar(nextChar)) {
-              skipNext = false;
-              skipPrev = false;
-            } else {
-              skipPrev = true;
-              skipNext = pattern.startsWith('*') && pos === 0;
-            }
-            return [j, skipNext, skipPrev];
-          }
-        }
-      }
-    }
-
-    return [defaultIndex, false, false];
-  }
-
-  /** Check if a character is in the polyphonic data */
-  private isPolyphonicChar(character: string): boolean {
-    if (!this.polyphonicData) return false;
-    if (!this.isChineseCharacter(character)) return false;
-    return character in this.polyphonicData.data;
-  }
-
-  /** Match a pattern against text around the character position */
-  private matchPattern(
-    pattern: string,
-    _pos: number,
-    start: number,
-    end: number,
-    text: string[],
-    character: string,
-  ): boolean {
-    let tmp = '';
-    for (let z = start; z < end; z++) {
-      tmp += text[z];
-    }
-    return tmp === pattern.replaceAll('*', character);
-  }
-
-  /** Check if a single character is a CJK unified ideograph */
-  private isChineseCharacter(char: string): boolean {
-    return CHINESE_CHAR_REGEX.test(char);
-  }
-
-  // ---------------------------------------------------------------------------
-  //  Style-set mapping
-  // ---------------------------------------------------------------------------
-
-  /**
-   * Map a matched variant index returned by match() to the BpmfIansui font's
-   * stylistic-set code ('0000' = no selector = font default, 'ss01'..'ss05' = variants).
-   *
-   * @param matchIndex  Index j returned by match(): -1 means "no explicit pattern
-   *                    matched", 0..n means v[j] was the matched variant.
-   * @param defaultVariantIdx  The v[] index whose pronunciation the font renders
-   *                    by default ('0000').  Stored as `d` in poyin_db.json;
-   *                    defaults to 0 when the field is absent.
-   *
-   * Mapping rules:
-   *   j == -1              → treated as j=0 (no match falls back to v[0] reading)
-   *   j == defaultVariantIdx → '0000'  (explicit match on the default variant)
-   *   j <  defaultVariantIdx → 'ss0{j+1}'  (shift up to skip the default slot)
-   *   j >  defaultVariantIdx → 'ss0{j}'    (no shift needed)
-   *
-   * When defaultVariantIdx == 0 (the common case), this simplifies to:
-   *   j == -1 or j == 0  → '0000'
-   *   j > 0              → 'ss0{j}'
-   * which is identical to the original code.
-   */
-  private variantIndexToStyleSet(matchIndex: number, defaultVariantIdx: number): string {
-    // No explicit match: fall back to v[0] as the reading for unrecognized contexts.
-    const j = matchIndex === -1 ? 0 : matchIndex;
-
-    if (j === defaultVariantIdx) {
-      return '0000';
-    }
-    if (j < defaultVariantIdx) {
-      return `ss0${j + 1}`;
-    }
-    // j > defaultVariantIdx
-    return `ss0${j}`;
   }
 
   // ---------------------------------------------------------------------------
@@ -504,7 +206,7 @@ export class PolyphonicProcessor {
       skipPrev = false;
 
       // Non-Chinese character: pass through with default style
-      if (!CHINESE_CHAR_REGEX.test(character)) {
+      if (!isChineseCharacter(character)) {
         result.push({ char: character, styleSet: '0000' });
         skipPrev = true;
         continue;
@@ -560,7 +262,7 @@ export class PolyphonicProcessor {
         // General tone sandhi for 一 and 不
         const prevTone = i > 0 ? getToneForChar(prevChar) : 0;
         const nextTone = i + 1 < length ? getToneForChar(nextChar) : 0;
-        const [newSs, skipNext, newSkipPrev] = this.getNewToneForYiBu(
+        const [newSs, skipNext, newSkipPrev] = getNewToneForYiBu(
           prevChar || null,
           character,
           nextChar || null,
@@ -649,8 +351,8 @@ export class PolyphonicProcessor {
           const variations = charData.v;
           if (variations && variations.length > 0) {
             const patterns = variations.map(String);
-            const [matchIndex, skipNext, matchSkipPrev] = this.match(
-              character, i, patterns, characters, skipPrevTemp,
+            const [matchIndex, skipNext, matchSkipPrev] = matchPolyphonicPattern(
+              character, i, patterns, characters, skipPrevTemp, this.polyphonicData,
             );
             skipPrev = matchSkipPrev;
 
@@ -658,7 +360,7 @@ export class PolyphonicProcessor {
             // Stored as charData.d; falls back to 0 for characters where v[0]
             // is the font default (the vast majority of polyphonic chars).
             const defaultVariantIdx = charData.d ?? 0;
-            const newSs = this.variantIndexToStyleSet(matchIndex, defaultVariantIdx);
+            const newSs = variantIndexToStyleSet(matchIndex, defaultVariantIdx);
 
             result.push({ char: character, styleSet: newSs });
 
@@ -685,23 +387,4 @@ export class PolyphonicProcessor {
 
     return result;
   }
-}
-
-// ---------------------------------------------------------------------------
-//  Utility functions
-// ---------------------------------------------------------------------------
-
-/**
- * Convert ProcessedChar[] to a Unicode string suitable for BpmfIansui font rendering.
- * Characters with non-default style sets get a PUA variant selector appended.
- */
-export function buildZhuyinString(processed: ProcessedChar[]): string {
-  let result = '';
-  for (const { char, styleSet } of processed) {
-    result += char;
-    if (styleSet !== '0000' && styleSet in SS_MAPPING) {
-      result += String.fromCodePoint(parseInt(SS_MAPPING[styleSet], 16));
-    }
-  }
-  return result;
 }
